requirements.txt
streamlit
google-generativeai
PyPDF2
langchain
langchain-community
langchain-google-genai
faiss-cpu